# deployment/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Book Lending API",
    "description": "ASP.NET Core Web API for managing book lending operations",
    "version": "1.0.0",
    "contact": {
      "name": "Book Lending App",
      "email": "support@booklending.com"
    }
  },
  "servers": [
    {
      "url": "https://api.booklending.com",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/books": {
      "get": {
        "summary": "Get all books",
        "description": "Retrieve a list of all books in the library",
        "operationId": "getBooks",
        "tags": ["Books"],
        "responses": {
          "200": {
            "description": "Books retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooksResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new book",
        "description": "Add a new book to the library",
        "operationId": "createBook",
        "tags": ["Books"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Book created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error or duplicate ISBN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{id}": {
      "get": {
        "summary": "Get book by ID",
        "description": "Retrieve a specific book by its ID",
        "operationId": "getBookById",
        "tags": ["Books"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Book ID",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          },
          "404": {
            "description": "Book not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{id}/checkout": {
      "post": {
        "summary": "Check out a book",
        "description": "Check out a book from the library",
        "operationId": "checkOutBook",
        "tags": ["Book Operations"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Book ID",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book checked out successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          },
          "400": {
            "description": "Book not available for checkout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{id}/return": {
      "post": {
        "summary": "Return a book",
        "description": "Return a checked out book to the library",
        "operationId": "returnBook",
        "tags": ["Book Operations"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Book ID",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          },
          "400": {
            "description": "Book was not checked out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Book": {
        "type": "object",
        "required": ["id", "title", "author", "isbn", "publisher", "pages", "unitsAvailable", "createdDate"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "Unique identifier for the book",
            "example": 1
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "description": "Title of the book",
            "example": "The Great Gatsby"
          },
          "author": {
            "type": "string",
            "maxLength": 100,
            "description": "Author of the book",
            "example": "F. Scott Fitzgerald"
          },
          "isbn": {
            "type": "string",
            "maxLength": 20,
            "description": "ISBN of the book",
            "example": "978-0-7432-7356-5"
          },
          "publisher": {
            "type": "string",
            "maxLength": 100,
            "description": "Publisher of the book",
            "example": "Scribner"
          },
          "pages": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "description": "Number of pages in the book",
            "example": 180
          },
          "unitsAvailable": {
            "type": "integer",
            "format": "int32",
            "minimum": 0,
            "description": "Number of units available for checkout",
            "example": 5
          },
          "checkedOutDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Date when the book was checked out (null if not checked out)",
            "example": "2024-10-06T10:30:00Z"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "description": "Date when the book was added to the system",
            "example": "2024-01-15T09:00:00Z"
          }
        }
      },
      "BookRequest": {
        "type": "object",
        "required": ["title", "author", "isbn", "publisher", "pages", "unitsAvailable"],
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "description": "Title of the book",
            "example": "The Great Gatsby"
          },
          "author": {
            "type": "string",
            "maxLength": 100,
            "description": "Author of the book",
            "example": "F. Scott Fitzgerald"
          },
          "isbn": {
            "type": "string",
            "maxLength": 20,
            "description": "ISBN of the book (must be unique)",
            "example": "978-0-7432-7356-5"
          },
          "publisher": {
            "type": "string",
            "maxLength": 100,
            "description": "Publisher of the book",
            "example": "Scribner"
          },
          "pages": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "description": "Number of pages in the book",
            "example": 180
          },
          "unitsAvailable": {
            "type": "integer",
            "format": "int32",
            "minimum": 0,
            "description": "Number of units available for checkout",
            "example": 5
          }
        }
      },
      "BookResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Book retrieved successfully"
          },
          "data": {
            "$ref": "#/components/schemas/Book"
          },
          "error": {
            "type": "string",
            "nullable": true,
            "description": "Error message (null on success)",
            "example": null
          }
        }
      },
      "BooksResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Books retrieved successfully"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Book"
            }
          },
          "error": {
            "type": "string",
            "nullable": true,
            "description": "Error message (null on success)",
            "example": null
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "Failed to create book"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "description": "Response data (null on error)",
            "example": null
          },
          "error": {
            "type": "string",
            "description": "Detailed error information",
            "example": "ISBN already exists"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Books",
      "description": "Operations related to book management"
    },
    {
      "name": "Book Operations",
      "description": "Operations for checking out and returning books"
    }
  ]
}